Move stack command to MonitorManager
